+++ R/sienaGOF.r
@@ -19,12 +19,14 @@ test_gof <- function(object, ...) UseMethod("test_gof", object)
 test_gof.sienaFit <- function(object, auxiliaryFunction,
 		period=NULL, verbose=FALSE, join=TRUE, twoTailed=FALSE,
 		cluster=NULL, robust=FALSE,
-		groupName="Data1", varName, tested=NULL, iterations=NULL,
+		groupName="Data1", varName, tested=FALSE, iterations=NULL,
 		giveNAWarning=TRUE, ...)
 {
-	sienaGOF(object, auxiliaryFunction, period, verbose, join, twoTailed,
-		cluster, robust, groupName, varName, tested, iterations,
-		giveNAWarning, ...)
+	sienaGOF(object=object, auxiliaryFunction=auxiliaryFunction,
+	period=period, verbose=verbose, join=join, twoTailed=twoTailed,
+		cluster=cluster, robust=robust,
+		groupName=groupName, varName=varName, tested=tested,
+		iterations=iterations, giveNAWarning=giveNAWarning, ...)
 }
 
 ##@sienaGOF siena07 Does test for goodness of fit
@@ -32,7 +34,7 @@ sienaGOF <- function(
 		object,	auxiliaryFunction,
 		period=NULL, verbose=FALSE, join=TRUE, twoTailed=FALSE,
 		cluster=NULL, robust=FALSE,
-		groupName="Data1", varName, tested=NULL, iterations=NULL,
+		groupName="Data1", varName, tested=FALSE, iterations=NULL,
 		giveNAWarning=TRUE, ...)
 	{
 	## require(MASS)
@@ -665,8 +667,6 @@ summary.sienaGOF <- function(object, ...) {
 		}
 		cat("\n-----")
 	}
-	cat("\nComputation time for auxiliary statistic calculations on simulations: ",
-			attr(x, "simTime")["elapsed"] , "seconds.\n")
 	invisible(x)
 }
 
@@ -1601,7 +1601,7 @@ egoAlterCombi <- function (i, obsData, sims, period, groupName, varName,
      trafo=NULL)
 {
 # An auxiliary function calculating the number of ties
-# for each ego-alter combination of values of the dependent variable;
+# for each ego-alter combination of values of the dependent behavior variable;
 # the dependent variable is transformed by trafo.
 	if (length(varName) != 2){
 		stop("egoAlterCombi expects two varName parameters")
@@ -1638,3 +1638,33 @@ egoAlterCombi <- function (i, obsData, sims, period, groupName, varName,
 	teax
 }
 
+egoAlterCovarComb <- function (i, obsData, sims, period, groupName, varName,
+     covar)
+{
+# An auxiliary function calculating the number of ties
+# for each ego-alter combination of values of covar.
+	varName1 <- varName[1]
+	varName2 <- varName[2]
+	m <- sparseMatrixExtraction(i, obsData, sims, period, groupName,
+		varName1)
+	m <- m[!is.na(covar), !is.na(covar)]
+	x <- covar[!is.na(covar)]
+	branges <- range(covar, na.rm = TRUE)
+	brange <- branges[1]:branges[2]
+	combi.egoalter <- outer(10*x, x ,'+')
+	possible.pairs <-
+		sort(unique(as.vector(outer(10*brange, brange, '+'))))
+	if ((length(possible.pairs) >= 100) && is.null(i))
+	{
+	warning("The variables used for egoAlterCovarComb has too many values.")
+	}
+	tmeax <- table((m * combi.egoalter)@x, useNA = "no")
+	ppnames <- as.character(possible.pairs)
+	teax <- setNames(0*possible.pairs, ppnames)
+	teax[dimnames(tmeax)[[1]]] <- tmeax
+	# pad names with leading 0s, if necessary:
+	pp.names <- ifelse(nchar(ppnames)==1, paste("0",ppnames,sep=""),ppnames)
+	names(teax) <- pp.names
+	attr(teax, "EgoAlter") <- TRUE
+	teax
+}