+++ src/model/effects/AllEffects.h
@@ -56,6 +56,7 @@
 #include "CovariateSimilarityEffect.h"
 #include "CovariateTransitiveTripletsEffect.h"
 #include "CrossCovariateActivityEffect.h"
+#include "DegreeMixedPopularityEffect.h"
 #include "DenseTriadsBehaviorEffect.h"
 #include "DenseTriadsEffect.h"
 #include "DenseTriadsSimilarityEffect.h"